requirements.txt
smolagents
smolagents[toolkit, telemetry]
pydantic
python-dotenv
openai
tavily-python